Update README.md
PowerBI module end Project
Here is the dataset of a store, which has branches all over the world. The stakeholders require the detailed report on their sale transaction.
Perform the basic data cleaning
Segment the visuals using country, region, market
Analyze the data by visualizing the percentage of shipping based on shipmode.
Give the report to see sales over the city, states, region and market.
Create tables for all the visualizations.
Make these visualizations and create an interactive sales report to show the stakeholders.
Create a video explaining the visualizations.